Improve white-space in code generated from MXML
This eases readability and debugging.
A tiny semantic change regarding MXML is that now all
white-space before a tag start ("<") is considered white-space
for that tag, so be careful with string literal values ending
with white-space. (Maybe they have been trimmed before?)
To declare a string with trailing white-space, use
  <fx:String>{"foo  "}</fx:String>
instead of
  <fx:String>foo   </fx:String>

This change was necessary to avoid white-space-only text nodes
before tags, but instead add the white-space to the tag.
This again leads to better indentation of field declarations
generated from <fx:Declarations>.