WIP: "New" MXML output format: Inline event handlers
Instead of generating methods, event handler code is now inlined again.
This again keeps generated code closer to the original source code.
By using scope: this, it is ensured that the event handler runs in the
same scope as if it were a method. The only thing that can still go
wrong is that the compiler currently cannot derive the type of "this"
in an anonymous function, and thus may not detect accessor usages
in some cases. To fix this, the compiler would have to learn about the
"scope" property.