Fix JOO-98: Parallel Maven builds may fail https://jangaroo.myjetbrains.com/youtrack/issue/JOO-98
Maven creates one instance for each execution of a goal, but the lock
was not static, so different instances would not be mutually excluded.

Also, the return value of setLastModified() was not checked, so such
file system problems would have gone unnoticed.

On the fly, changed error handling during new Sencha Cmd version
detection to only log a warning and assuming the old version instead
of throwing a build exception. This is safer to not break the build
for existing customers who use the recommended old version.