Fix JOO-106: Private field in subclass overrides MXML access of public field
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-106

As suggested in the evaluation, changing the type of the super config
object from current class to super class resolves the problem.
The config object having the wrong type also was the reason that in
MxmlToModelParser#findPropertyModel(), property lookup started in the
super class of the given class rather than looking in the given class
first. Now that the type is corrected, this lookup can be done in
a straight-forward manner. While at it, replaced recursion by a simple
loop.

All expected JS code generated from MXML had to be adjusted accordingly.