More Convenience for exml-to-mxml Migration Tool
Instead of downloading the ext-as artifact (using mvn dependency:get)
and then specifying the full path in the local Maven repository,
you now just have to only specify the ext-as version:
  mvn net.jangaroo:exml-maven-plugin:2.0.19:exml-to-mxml -DextAsVersion=...

The migration tool now also automatically takes care of changes in
Jangaroo 4 packaging and dependency types. If the ext-as version is
6.2.0-13 or higher, packaging "swc" is chosen instead of "jangaroo-pkg",
and in the POMs of all Jangaroo modules, the dependency type "swc" is
added to all dependencies.