Fix Generating App from Template with Cmd 6.5
With Cmd 6.5, you also need to set the new flag --full for
"generate app", not only for "generate workspace".
Although for "generate app", this is undocumented, Sencha support
advised us to use the flag --full to make our custom application
templates work like before (see ticket #40966).

To reuse the functionality "add --full flag if Cmd version >= 6.5",
extracted it into a utility method. Note that this method throws an
exception if the Sencha Cmd version cannot be determined, unlike
before, where this was logged as a warning and assumed that Cmd
version is < 6.5. In the meantime, the detection has been improved
and the assumption is more likely false, so throwing an exception
seems the better choice now.