Fixes JOO-118: Compiler does not detect certain undeclared identifiers
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-118

The compiler should now report all undeclared identifiers. To achieve
this, any identifier expression (IdeExpr) that is *not* rewritten to a
dot expression is checked to have a declaration, else an error is logged.

A new test Jooc#testUndeclaredIdentifier() with a test class
TestUndeclaredIdeError.as ensures that such errors have not been found
before (red without the fix in IdeExpr.java) and are now found (green
when including the fix in IdeExpr.java).

Also, the fix revealed an error in another test class,
TestOldPropertyAccessSyntax.mxml, which is not even detected by IDEA:
When inheriting from Object (top-level <fx:Object> element), although
Object is a dynamic class, the subclass does not inherit being dynamic,
so setting an undeclared property "foo" now correctly leads to a
compile error. To fix the test code, it now declares a new property
"foo" with a default value. The expected transpiled JS output has been
adapted accordingly.