Fixes JOO-119: Compiler does not detect incorrect super calls
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-119

The compiler should now report all incorrect super calls. The type
checker for this is quite straight-forward, the real problem was that
method signatures were assigned in analyze phase, where they are
already needed for type checking.

To be able to access method signatures in analyze phase, they have to
be assigned in scope phase. Here, the problem was that creating a
FunctionSignature lead to creating ExpressionType objects for all
parameter types and the return type, which in turn tried to resolve
the referenced types. This may fail in scope phase, because not all
declaration have yet been scope and thus some cannot be resolved.

The solution was to make type declaration lookup lazy in
ExpressionType.java and AbstractScope.java.
Then, assigning the method signature could be moved from analyze to
scope phase (FunctionExpr.java).
Eventually, TypeChecker could easily be extended by a check for super
constructor calls (visitSuperConstructorCallStatement()) by extracting
the parameter check code from visitApplyExpr() into an auxiliary
method (checkParameterTypes()).