[ext-as-api-generator] Spike: Generating Ext AS API from Doxi Output
Sencha has switched from jsduck to a new documentation tool called
Doxi, which is integrated in Sencha Cmd. Using this "hidden feature",
you can generate a JSON file describing the complete Ext JS (Classic)
API using

    sencha doxi build -p [doxi project file name]

Unfortunately, the Doxi project file is not part of the Ext SDK
download, but Sencha kindly gave me access to the repo that holds
this file.

ExtJsApi and ExtAsApiGenerator are now revamped to deal with Doxi's
JSON format instead of JsDuck's. This is work-in-progress,
get/set methods for accessors are still missing and many things have
to be cleaned up.
We use the Java library FlexMark to convert Markdown to HTML (ASDoc
does not understand Markdown).

On the fly, switched ext-as-api-generator to Java 8 language level.