[ext-as-api-generator] Convert @link to @see
Instead of {@link ref text},
use <em||code>text || ref</em||code> + @see ref text

This is not as easy as it sounds, since
* sometimes, {@link} is already enclosed in `..` or **...**
* the link text is often redundant and really a code reference which
  should be translated to ActionScript, too
* event classes and interfaces need special treatment so that the
  ASDoc references are valid (see thisClassName)
* ASDoc method references need parentheses ("()"), JSDoc references
  (sometimes) use #method!foo or #method-foo or #foo()
* sometimes the link text contains additional information, e.g. the
  member type "method" or "()" which is lacking in the actual link

Thus, we have to normalize the link and the link text first, merge them,
compare them and if the link text is just a rephrase of the link, leave
out the link text in @see and rewrite it for the inline text.
This is sufficiently complicated to be moved to a new separate (inner)
class JsDocReference.