[ext-as-api-generator] Resolve JSDoc references without member type
In Ext JS 6.5.3, far more JSDoc references come without member type
information. Thus, we do not know wheter #foo is a property, a method,
or an event. Since valid ASDoc depends on this information, we now
resolve the member to find out its member type.
The ugly hand-through of "thisJsClassName" could be improved.