[ext-as-api-generator] Document events only once
ASDoc duplicated event documentation, because when using @eventType
in an [Event] annotation, ASDoc fetches the documentation from the
references event name constant. Since IDEA does *not* do this, we
had put the same documentation at the [Event] and at the constant.
Now, the [Event] only uses an @see (because IDEA fails to make the
@eventType navigable :-() and the actual documentation is only found
at the event name constant in the event class.
Also, let ActionScriptCodeGeneratingModelVisitor move @eventType
together with @see, as the output looks weird (at least in IDEA)
when @see comes after @eventType.