[ext-as-api-generator] Fix {@link\n#...}
Fix rare cases were there is a new line inside an {@link} followed by
a hash, which is then interpreted as a headline by the Markdown parser,
because it is positioned at the beginning of a line.
Trying to process @link-s before transforming to Markdown got too
complicated because that processing checks for <code> and <em> regions,
which look different in Markdown, which would have required quite a
rewrite.