[ext-as-api-generator] Inline technical link syntax and simplify references to singletons
Instead of "footnote style", we now inline all @link-s that have no
link text or where the link text is just a shorthand form of the
technical link syntax.
The @see entries only receive a link text for references to singletons
to hide the singleton type, which is just a helper construct for the
ActionScript type system.
A bit code clean-up is to follow.