[ext-as-api-generator] Use '.' instead of '#', suppress leading '#'
Actually, when rendering a "technical" link, asdoc replaces '#' by '.'
and suppresses a leading '#'. Unfortunately, IDEA does not do that
but leaves technical links as-is, but we hope that developers will
still recognize ->Foo.bar() and "See Foo#bar()" as the same thing.
Anyway, leaving out the '#' increases readability and is consistent
with asdoc.