[ext-as-api-generator] Render sub-params / properties as ul/li
ASDoc tool cannot cope with @param-s with dotted names. Thus, removed
the code to render @param tags for sub-properties and always use the
code that renders <ul> with <li> elements.
Replaced the paramPrefix by a simple flag to still know the difference
between top-level properties (for which a reference to the Sencha doc
is generated) and params and sub-properties (for which no such link
must be generated).