Fix JOO-125: Good code red: Array literal incorrectly parsed as annotation
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-125

For details, see issue description.

Thanks to agawecki, this could be fixed by simplifying the grammar.
LBRACK_EXPR was introduced analogously to LBRACE_EXPR to disambiguate
the usage of '[' between expressions (array literal) and annotations.
When removing the auxiliary token LBRACE_EXPR, the parser generator
reports a new shift/reduce conflict between annotation ::= '[' ide
and qualifiedIde ::= ide, but since this conflict is resolved in favor
of shifting, we end up with the desired behavior. All we have to do is
allow this additional conflict (the other one is dangling else) in the
Maven plugin configuration in pom.xml.