Working on JOO-129: Fix executing unit tests with global resources
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-129

After changing how unit test apps are built and how global resources
are mapped, unit tests with global resources failed because the
generated code tried to use the new helper function
Ext.registerGlobalResources(), which is defined in before-ext-load.js
and not (yet) loaded in test apps.
For the time being, copied the utility function to the test app
index.html (ugly) and complemented default.test.app.json to initialize
globalResources to an empty object.