Working on JOO-129: Use jangaroo:run for local module development
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-129

Now that modules can be loaded at run-time, a convenient developer
setup is to reuse an app from a reference system and "mix in" an
additional module. The reference app contains a dynamic-packages.json
as a "module inventory", so the local module has to be added to this
resource. For that purpose, a simple servlet is introduced that reads
dynamic-packages.json from the proxied app, adds the name of the local
module, and serves the complemented JSON array as the response.
In module mode, jangaroo:run now proxies *everything* (REST requests
*and* static resources) from/to the reference system, blends in the
local module code under ./packages/<local-package-name>, and adds its
package name to dynamic-packages.json as described above.
Thus, you end up with a server that is exactly like the reference app
plus your local dynamic package module.