Repackage Ext JS Charts
As for ext-all, we want to aggregate all JS code from the charts
package into one (or few) file(s) and use a sourcemap for debugging.
This is achieved by moving the actual sources (src and override
subfolders, also under classic subfolder) and placing the aggregated
sources as one file in the src folder where Sencha Cmd looks for the
sources.
Unfortunately, one (!) charts source file uses a compiler directive
// @define ...
instead of Ext.define(), and Google Closure Compiler removed all inline
comments. The solution is to scan all source files for // @define
directives and add these at the end of the aggregated file, but before
the //#sourceMappingUrl directive (which has to be last).
For consistency, renamed the directory for source-mapped sources in
themes from overrides-src to sourcemap.