Fixes JOO-136: Jangaroo app overlay dependency resolution ignores remote repositories declared in POMs
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-136

The default artifact lookup scope for project building requests contains
the remote repositories from the Maven session - which are actually the
remote repositories that are defined in the _settings.xml_. It doesn't
include remote repositories that are defined in the Maven project POMs
and are not cover by any configured mirror. To fix the lookup scope, the
remote repositories of the project building requests must be overwritten
with the projects remote artifact repositories - configured mirrors are
properly covered by them.