Fix JOO-138: App Overlay adds too many packages to dynamic-packages.json
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-138

Refactored overlay artifact computation. Now, even third level overlays
should work. A helper class represents a JangarooApp or, as a special
case (subclass) a JangarooAppOverlay. An App Overlay refers to its base
app, which may again be an App Overlay or (termination) an App.
For each level, transitive dependencies are collected
(PreparePackageAppOverlayMojo#populatePackages()), so that its "own"
(exclusive) packages can easily be computed.

RunMojo (jangaroo:run) has been adapted to use the new helper methods.
On the fly, introduced a constant for "/*" in JettyWrapper and fixed
a corner case in RunMojo / JettyWrapper where both the proxy servlet
and the default (static resources) servlet ended up to be mapped to
the root path ("/*"). Now, the default servlet is added for the root
path exactly if no other servlet occupies this path.