Speed up by reducing number of ProjectBuildingRequests
All usages of AbstractSenchaMojo#createProjectFromDependency() have been
replaced by #getProjectFromDependency(), which takes the project
containing the dependency as an additional parameter. This optimizes
the case where a dependent project is also contained in the current
Maven Reactor (the project's "project references"), so that some of the
rather expensive ProjectBuildingRequests are avoided.