Fix JOO-144: Update HTTP Proxy Used by Jangaroo Run
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-144

This commit is based on changes by mmichaelis. Thanks!

Updated D. Smiley's HttpProxyServlet from 1.9 to 1.11.
For that, code in the patching subclass JangarooProxyServlet has to
be adapted.
Improved documentation of what the actual patch is.

Refactored HttpClientUtil to only create the specific
SSLConnectionSocketFactory (which accepts any certificate) instead
of using HttpClientBuilder.