JOO-145 jangaroo:run incorrectly proxies URLs containing escaped charaters
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-145

Indeed, getPathInfo() is specified to return a decoded value, while
getRequestURI() returns a non-decoded value. Consequently, when
emulating getPathInfo() through getRequestURI(), we have to decode
the URI, even before subtracting the servletPath (which is also
already decoded).
Hopefully, URLDecoder.decode() with UTF-8 encoding is the correct
decoder here.
Added an assertion that the requestURI actually starts with servletPath.