Improved outer this access solution using arrow functions
When there is no access of dynamic "this", ActionScript inner
functions can be replaced by JavaScript arrow functions, which
then automatically access the outer "this".
We now always convert functions to arrow functions if possible.
This produces much more elegant code.
The only downside is that with the current approach,
"const this$=this;" cannot be suppressed, because we need to
complete "analyze" phase to know which functions can be converted
to arrow functions. For arrow functions, we generate "this" instead
of "this$", but do not know whether any usages of "this$" (rewriteThis)
remain before generating its assignment. I hope we can optimize away
the assignment later, or let IDEA do that.