[ExtAsApiGenerator] Ensure AbstractPlugin and Mixin subclasses generate mixins
Unfortunately, in theory, any Ext class can be used as a mixin, so
ExtAsApiGenerator makes all classes mixins that are used as a mixin
within the framework. However, some classes are mixins only by
documentation and may not be used inside the framework.
Ext.plugin.Abstract used to be mixed into some Ext.ux class, but with
Ext JS 7.1, this usage is gone and thus ExtAsApiGenerator no longer
recognizes it as a mixin class.
The code change makes this case explicit, but we should find another
way to determine which Ext classes are mixins. Being a subclass of
Ext.Mixin is of course another sign for a mixin (currently, only
direct subclasses are detected).