Interfaces as abstract classes approach
Surprisingly, in TypeScript, an abstract class implementing
an interface must "implement" all interface methods (but
the actual implementation may be missing). To prevent this
duplicate declaration, somebody found a solution that takes
advantage of declaration merging:
https://github.com/Microsoft/TypeScript/issues/4670#issuecomment-326585615
So we now convert an ActionScript interface into a TypeScript
interface *plus* a TypeScript abstract class of the same name.
Extended interfaces are extended by the interface and
"implemented" by the abstract class.
This keeps a runtime presentation of the ActionScript interface
that can be translated to Ext JS mixins and used for "is" checks.