Remove unused code to render imports
It should never be called, since ActionScript imports
are just used to resolve compilation unit references,
and TypeScript imports are rendered independent of these.
We _could_ try to keep the order of AS import statements
in TS, but that could also better be handled centrally
than for each single import directive.