Hide private member declarations
Like in ActionScript and (through a trick) in Jangaroo,
private members are now hidden from subclasses. Instead
of renaming them using a fqn-hash suffix, we use ECMAScript
Symbols as unique slots. For each private member, a
module-private Symbol constant is defined, named like the
member plus "$", and the Symbol value is also the private
member name to ease debugging. The constant is used to
declare and access the private members.
This is not yet implemented for private static members,
since it might be easier and more intuitive to define
them on module level than inside the class.
To allow private statics to access private members, these
would have to be no longer TypeScript-private, but that
does not make any difference now, anyway, since the private
member symbols are not accessible from outside the class.