Fix for-each loops for non-array types
When the expression a for-each loop iterates over
is *not* an array-like type (not "iterable"), to
achieve the corresponding ActionScript semantics,
we must use Object.values() on the expression to
use it in ECMAScript's for-of loop.
We could even do this for Arrays, but it would
be redundant and thus inefficient.