Fix: constructor: no return type
In ActionScript, a constructor usually does not have a
return type, but if you add one, it is simply ignored
(at least in IDEA).
In TypeScript, a constructor must not have a return type,
this is an error.
FlexUnit has some ActionScript classes where constructors
return `void`. This must be suppressed for TypeScript.