New interface transpilation
Instead of interface + abstract class, an ActionScript
interface is now transpiled to a TypeScript interface
plus a run-time representation of it, which can be applied
as a marker to classes to perform "is" checks.
This works very well and even revealed a bug in AS code,
where "instanceof" was used with an interface, which always
returns false.