Fix: suppress new-lines between "readonly" and identifier
In a class field declaration, a new-line between the "readonly"
syntactic keyword and the variable identifier leads to
TypeScript interpreting "readonly" as the identifier and
auto-adding a semicolon, so this does not even lead to a
syntax error.
Although a rare case, we better replace all white-space between
readonly and identifier by a single space.