Transpile native AS methods correctly into TS
"native" AS methods now generate valid TS.
Most "native" AS methods are accessors. Accessors are
the only way to declare properties in AS interfaces.
Unfortunately, AS does not allow to implement a pair
of get+set interface methods simply as a public var,
and a sole get accessor as a public const. To support
this without changing AS syntax, Jangaroo allow "native"
accessors in implementing classes and transpiles them
to JavaScript fields. This is very similar to how
TypeScript treats properties in interfaces: In TS,
interfaces may define properties directly, consequently,
accessors are not allowed. For "const" properties,
a "readonly" keyword is used. (You cannot define
"writeonly" properties in a TS interface, but there is
an open feature request for that.) An interface
property can be implemented either by get+set
or just get ("readonly"), or by a var or const,
respectively.
All this means that AS "native" accessors simply become
class fields in TS, "readonly" if there is only a getter.

All other "native" non-accessor methods are now
moved to a "companion interface", which is an
interface with the same name as the class.
TypeScript performs "declaration merging",
resulting in the type aspect of the class being
extended by the interface without generating any
additional JavaScript code.
TODO: If native methods implement Mixin interface
methods and contribute no documentation, they can
be suppressed, and instead, the mixin interface
would be extended.