Restore standard JS output mode
Hide the TypeScript migration behind a flag `migrateToTypeScript`.
This flag is now present in Maven, as a command-line flag `-ts`
(`--migrate-to-typescript`) and in the compiler API.
Some code that is actually not affected by or are (yet) not
supposed to support TypeScript migration (Ant task,
MergedOutputCompilationUnitSinkFactory, tests) have been reverted
to their master state.
All other changed code now respects the new flag and (hopefully)
behaves as before if the flag is not set. At least all tests
are now green again.
The flag makes it much easier to keep JS and TS output mode
consistent and to switch between them.