Move call field initializer code to super constructor call method
Since this is generated code and in almost every case, we have a
super constructor call method, anyway, it makes sense to move
call field initializer code to this super constructor call method.
It turned out a bit tricky for helper classes, but it seems to
work fine now.
The only minimal overhead is that classes directly inheriting
from Object now also have a super$xyz method for calling the
field initializers, which were inline in the constructor before.