Generate super() _after_ parameter initializers
If super() is missing in an ActionScript constructor,
an empty super() call is implicitly added as the first
command.
Logically, parameter defaults must be assigned even before
making a super constructor call. But Jangaroo added the
implicit super call before parameter init code.
Actually, this cannot make a semantic difference, since
parameters are local, their initializers are constant
expressions that do not execute any code, and an empty
super() call obviously cannot use them.
Still, for code execution order, it is cleaner to always
generate parameter init code first, before the super() call.

On the fly, refactored to separate handling of parameter
initializers and the ...rest parameter.

Also, added a test for generated JS code of parameter
initializers, reusing the existing ParameterInitializers.as,
which was so far only ran through the API compiler.