Alternative emit of parameter initializers (disabled)
A by default disabled alternative implementation of
getParameterInitializerCodeGenerator(), which emits JS
code for ActionScript parameter initializers.
The flag to enable this alternative is still missing,
to try it out, one has to replace the (only) call to
getParameterInitializerCodeGenerator() by calling the
new method getEcmaParameterInitializerCodeGenerator().

In ActionScript (IIRC), parameter initializers only
kick in when the method is called with less arguments.
In the ECMAScript standard, a different semantics has
been decided: Any `undefined` parameter is assigned
its initializer value. So the semantics difference
only matters when calling a method with an explicitly
`undefined` argument for a parameter with initializer,
as in this example:

  function foo(param:string = "default") {
    return param;
  }
  foo("FOO"); // "FOO" in both cases
  foo(); // "default" in both cases
  foo(undefined); // undefined in AS3 semantics,
  // "default" in ECMAScript semantics

To mitigate migration problems ahead of time, we
should try if changing the semantics from AS3 to
ECMAScript really breaks anything in practice.
If not, we should do the switch *before* migrating
to TypeScript, so people can get used to this
behavior.