[TS] Relative and NPM-friendly module references
In JavaScript, when importing modules, source-root-relative
paths are totally unusual. You have to define a baseUrl in
TypeScript to make these even work.

Instead, relative paths are used to import modules. If the
imported module is from another NPM package, absolute paths
are used, starting with the NPM package name.

Up until now, only Jangaroo Maven Plugin needed to deal with
Sencha package names. Fortunately, the package name is
included in the SWC artifact of dependent modules. It is the
name of the *.json file under META-INF/pkg, excluding the
package.json that is also placed there.  To be able to access
the SWC contents, a getter had to be added to ZipEntryInputSource.

Note that currently, the JSON file is missing in an SWC built
with migrateToTypeScript switched on, so for successful migration,
you first have to 'mvn install' with migrateToTypeScript turned
off, then 'mvn clean test-compile' with migrateToTypeScript turned
on. As this is a pitfall, it must be fixed soon.

The current code implements a simple approach to convert the
Sencha package name to a (shorter) NPM package name. It maps
'net.jangaroo__' to '@jangaroo' (NPM "vendor prefix") and the
very special case 'jangaroo-runtime' to 'joo'.
The mapping should be configured through a Jangaroo Maven Plugin
parameter.

Another special case is test code that references "main" code
to test. This is the only case where in Jangaroo's Maven
build process, a compilation unit can come from a different
source root dir without using an SWC. We generate a relative
path that assumes that in the TypeScript workspace, tests
sources are placed in a sibling directory to the 'src' directory
where the "main" sources are located. The string "tests" in
method getRequireModuleName() is actually a dummy just to
get the right amount of '../'.