Suppress redundant parenthesis around !(AS3.is(foo, Foo))
Since (in JavaScript as well as TypeScript), the infix
operators 'is' and 'as' are translated to function
calls ('AS.is(foo, Foo)'), surrounding the expression
with additional parenthesis becomes redundant.
For TypeScript migration, suppressing the redundant
parenthesis makes comparison with code generated via
Babel easier, since Babel automatically suppresses
redundant parenthesis in JavaScript code. In most
other cases, you can remove the redundant parenthesis
in the AS3 source, but in this case, this was not
possible, so we let the compiler do it.

Implementation:
* pulled up visitInfixOpExpr() to CodeGeneratorBase
* let visitParenthesizedExpr() detect inner InfixOpExpr
  of OpExpr and reuse lParen and rParen symbols
* added test case to check that redundant parenthesis
  actually disappear