[AST] Fix: set type of array literal correctly
It was set before super call, which, in ParenthesizedExpr,
overwrote the 'Array' type by the sub-expression's type
(which would be a single element or a comma expression).

TODO: We could use the sub-expression's type to determine
the type parameter of the 'Array<T>'. This could at least
be done if all elements of the comma expression have the
same type ('["A", "B"]' would then be of type
'Array<String>').