Fix: Set implicit ...rest parameter type to Array
If a ...rest parameter has no type annotation, it
implicitly is of type Array.
AST class Parameter now synthesizes such a ': Array'
type relation. JsCodeGenerator takes care of that
"virtual" type relation not being generated (as
comment) as JavaScript code.

The newly considered type of ...rest parameters
revealed a bug in AbstractScope#getFunctionSignature().
Special handling of a ...rest parameters was done
_after_ adding it to the parameterTypes list, so
it was treated like a normal parameter of type Array,
causing incorrect type errors.

Two expected test code files had to be adjusted,
because for-loops generate different code when an
Array is detected.