Fix: Set implicit ...rest parameter type to Array
If a ...rest parameter has no type annotation, it
implicitly is of type Array.
AST class Parameter now synthesized such a ': Array'
type relation. CodeGeneratorBase takes care of that
synthesized type relation not being generated (as
comment) as JavaScript code and provides method
visitParameterTypeRelation() for subclasses to reuse
that logic, which both JsCodeGenerator and
TypeScriptCodeGenerator do.

The newly considered type of ...rest parameters
revealed a but in AbstractScope#getFunctionSignature().
Special handling of a ...rest parameters was done
_after_ adding it to the parameterTypes list, so
it was treated like a normal parameter of type Array,
causing incorrect type errors.

Two expected test code files had to be adjusted,
because for-loops generate different code when an
Array is detected.