Fix: NPM package name heuristic must ignore <package>-overrides.json
SWC not only contain META-INF/pkg/<package-name>.json, but also
META-INF/pkg/<package-name>-overrides.json. Be careful not to use
the wrong one!
The real solution would be to read and parse
META-INF/pkg/package.json to extract the top-level JSON property
'name', but parsing JSON is so far only done in the Maven Plugin
module, so I wanted to avoid another 3rd-party dependency in the
core compiler, which would also have to be added to the IDEA plugin.