[TS] Map AS3.Error correctly
To use the same [Native] annotation on top-level AS3 class
'Error' in AS3 and TS mode, namely
  [Native("AS3.Error", require)]
the TS code generator must know of this exception to avoid
the overhead of an artificial Error class in TypeScript.
In TypeScript, unlike in Ext JS, you can use the built-in
Error class as a superclass directly.
The Jangaroo Babel Plugin then maps TypeScript 'Error'
back to 'AS3.Error', which must be required.