[TS] Handle AS3 Object type and dynamic properties
'Object' and sqare-bracket property access are not the
same thing in ActionScript and TypeScript.

In AS3, the 'Object' class is defined with a 'dynamic'
keyword, which allows to access any non-declared
property of an object of type 'Object'.
Using square brackets with a string literal allows
dynamic, completely untyped access to properties
even of typed objects.

In TypeScript, neither type 'Object' nor 'object'
allow to access any non-declared properties. You
must use square bracket property access even then.
The only type that allows using dot property access
in an untyped fashion is 'any' (or 'unknown', which
is (almost?) the same thing).
Even when using square brackets, if the string literal
refers to a declared property of the object's type,
this property declaration is found and checked.

These different semantics lead to the following mapping
(AS3 -> TS, var obj:Object, var typed:SomeClass):
* obj.prop -> obj['prop']
* obj['prop'] -> obj['prop']
* typed.declaredProp -> typed.declaredProp
* typed['declaredProp'] -> typed['declaredProp']
  (typed.declaredProp would also work)
* typed['undeclaredProp'] -> (typed as object)['undeclaredProp']
  (typed as any).undeclaredProp would also work)

Since we now handle ArrayIndexExpr, anyway, on the fly,
insert a 'String(...)' conversion when the index expression
is not of type 'any', 'number' ('int', 'uint'), 'string'
(see TYPES_ALLOWED_AS_INDEX).