[jangaroo:run] Delete non-functional resource mappings for jangaroo-apps
The improved JettyWrapper resource mapping console output
revealed that two mappings created for jangaroo-apps did not
have any effect and can safely be removed. The root context
resource mapping already finds the resources the non-functional
configs tried to map, as they are in the "global" base resources.
The global resource mapping here is also redundant, because
JettyWrapper already takes care of adding this mapping if there
is no custom mapping for the root context.