[TS] Fix: without file extension, dirs may be mixed up with files
In ext-as, there is a directory 'animation' next to a file 'Animation.ts'.
Under Windows, when leaving out the file extension, new File() returns
the directory, even though the file is upper case.
To avoid this ambiguity, we add the ".ts" extension (actually, it could
be any file extension!) and remove is after resolving the relative path.