[MXML] Back to nested objects instead of aux vars
When compiling MXML to ActionScript, with Jangaroo 4,
we switched from generating nested objects to creating
auxiliary variables for each object and assigning
properties one-by-one. The reason was that this allows
stronger typing in ActionScript, which cannot type-check
object literals.
However, when migrating to TypeScript, this no longer
holds true. TypeScript is very good at type-checking
object literals. Nested objects look *much* cleaner and
much better resemble the nested MXML structure.
To use the same solution for JS and TS, we are going to
implement static type checking for object literals
(when "casting" an object literal to a type) in Jangaroo.

This commit is WIP of revamping the MXML compiler to
generate nested objects. So far, only one "expected"
JS file has been adapted, AllElements.js, and the test
still fails, but we're getting close.
Open issues are declarations (field assignments versus
applying on a "default" config object) and when to
instantiate configs immediately.