[MXML] Fix field assignment of elements with id
Declarations (field assignments versus applying on
a "default" config object) is now handled like before.
The case 'this.gear' is using an id attribute on a
config field, but not in <fx:declaration>, and was
never really supported. If would need to be assigned
as a top-level config default and use 'config.gear'
when used in a nested element, but that would require
a deep scan for all elements with an id attribute.
Fortunately, this usage is irrelevant for Studio and
thus can be ignored.