[MXML] Fix: also consider generated constructor parameter
If in an MXML file, no 'native' constructor is given, a default
constructor is generated. Since it always featured a 'config'
parameter, this parameter should be used as 'constructorParam'.
Not doing so caused weird generated code that would compute
config objects, but do nothing with them. Now, the only downside
when you forget to add a 'native' constructor is that this makes
it harder to create this MXML component from ActionScript code.
Since 'new MyMxmlClass()' does not seem to take a 'config'
parameter (in "real" Flex, MXML classes have a no-arg constructor),
you must use 'Ext.create(MyMxmlClass, { ... })'.