[MXML] Fix regression: Add fields for 'default' property
In Flex, a property can be annotated as [DefaultProperty].
Then, if you nest components directly into the component
with such a default property, this is treated as if the
inner components are the array value of the default property.
This is quite useful for Container#items, but since it hinders
type checking and completion in IDEA, we decided against
using the feature in Ext AS. Still, there is a test for it
and it may have been used out there, so it better work.