[MXML] Improve handling of elements with id and type guessing
Elements with 'id' are now always assigned to the corresponding
field if they are not directly nested inside <fx:Declarations>.
This is a rare case and supposedly never used at all in Studio,
but it occurs in test code (SimpleMxmlClass id="no_config").

Type guessing for primitive types is now activated in all
places (see test result improvements in SimpleMxmlClass.js).

There are now some more Exml.asString() calls, but it can be
argued that they should actually be there and at least they
do no harm, since the target property is of type String, so
they have been added in the expected test files.