[MXML] Generate type assertions for object literals (config objects)
Ext config objects must not always be complemented by 'xclass', 'alias'
etc., but sometimes just should be type-checked, either by Jangaroo or,
after migration, by TypeScript.
For that purpose, we introduce a "virtual" function __typeCheckObjectLiteral__
that is only used within the compiler.
The MXML compiler generates calls to __typeCheckObjectLiteral__ to
"hand over" type-checking to the AS3 compiler. After type-checking (not
yet implemented, see TODO in ApplyExpr.java!), the AS3->JS compiler simply
suppresses the virtual function call (in JsCodeGenerator), while the AS3->TS
compiler translates it to a call to 'AS3._<T1 & ... & Tn>(O)', which is
then suppressed by the Jangaroo Babel plugin later, but still valuable
for TypeScript type checking.

This solves the problem of typed, but xclass-less objects like 'defaults'
and values of the 'mixins' property in a type-safe way.
It also removes the runtime overhead of creating a config object with 'xclass'
etc. that is handed into a config-enabled constructor directly. The config
object is now only type-checked, the runtime conversion through 'AS3.cast()'
is no longer generated in such cases. You find examples of this change in
the changed expected test files.